split megafile contents around a little
